backend/restaurant-crawling/scripts/requirements.txt
google-api-python-client
python-dotenv
yt-dlp
pyyaml
openai
requests
langchain
langchain-community
langchain-ollama
supabase
